Notify immediately after download
