+++ opensrp-chw/src/main/java/org/smartregister/chw/application/ChwApplication.java
@@ -157,8 +157,8 @@ public class ChwApplication extends CoreChwApplication {
     private static void createFolders(String rootFolder, boolean onSdCard) {
         try {
             FileUtils.createDirectory(rootFolder, onSdCard);
-            FileUtils.createDirectory(rootFolder + File.separator + "en", onSdCard);
-            FileUtils.createDirectory(rootFolder + File.separator + "fr", onSdCard);
+            //FileUtils.createDirectory(rootFolder + File.separator + "en", onSdCard);
+            //FileUtils.createDirectory(rootFolder + File.separator + "fr", onSdCard);
         } catch (Exception e) {
             Timber.e(e);
         }
