+++ opensrp-chw/src/main/java/org/smartregister/chw/util/DownloadGuideBooksUtils.java
@@ -11,8 +11,6 @@ import java.io.File;
 
 public class DownloadGuideBooksUtils extends DownloadUtil {
 
-    private GuideBooksFragmentContract.DownloadListener downloadListener;
-
     public DownloadGuideBooksUtils(GuideBooksFragmentContract.DownloadListener downloadListener, String fileName, Context context) {
         this.fileName = fileName;
         folder = Environment.getExternalStorageDirectory() + File.separator +
@@ -20,20 +18,5 @@ public class DownloadGuideBooksUtils extends DownloadUtil {
                 context.getResources().getConfiguration().locale + "/";
         this.downloadListener = downloadListener;
         this.serverUrl = BuildConfig.guidebooks_url + context.getResources().getConfiguration().locale + "/" + fileName;
-    }
-
-    /**
-     * Before starting background thread
-     */
-    @Override
-    protected void onPreExecute() {
-        super.onPreExecute();
-        downloadListener.onStarted();
-    }
-
-    @Override
-    protected void onPostExecute(String message) {
-        // dismiss the dialog after the file was downloaded
-        downloadListener.onDownloadComplete(!"Something went wrong".equals(message));
     }
 }
